Echo passed arguments
remove the printed help every time the function is called
refactor the gym-dynamic pong test files
